pyproject.toml
[project]
name = "github-linkedin-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel",
    "pygithub",
    "langchain-community",
    "langchain-google-genai", 
    "langgraph",
    "python-dotenv",
    "mcp[cli]",
    "httpx"
]
